Move `parsePath` to `lib/fs`
