+++ lib/fs.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const fsp = require('node:fs').promises;
+const path = require('node:path');
 const { toBool } = require('./async.js');
 
 const directoryExists = async (path) => {
@@ -15,7 +16,15 @@ const ensureDirectory = async (path) => {
   return fsp.mkdir(path).then(...toBool);
 };
 
+const parsePath = (relPath) => {
+  const name = path.basename(relPath, '.js');
+  const names = relPath.split(path.sep);
+  names[names.length - 1] = name;
+  return names;
+};
+
 module.exports = {
   directoryExists,
   ensureDirectory,
+  parsePath,
 };
