+++ lib/strings.js
@@ -1,7 +1,5 @@
 'use strict';
 
-const path = require('node:path');
-
 const replace = (str, substr, newstr) => {
   if (substr === '') return str;
   let src = str;
@@ -66,13 +64,6 @@ const fileExt = (fileName) => {
   return fileName.substring(dot + 1, fileName.length).toLowerCase();
 };
 
-const parsePath = (relPath) => {
-  const name = path.basename(relPath, '.js');
-  const names = relPath.split(path.sep);
-  names[names.length - 1] = name;
-  return names;
-};
-
 const trimLines = (s) => {
   const chunks = s.split('\n').map((d) => d.trim());
   return chunks.filter((d) => d !== '').join('\n');
@@ -93,6 +84,5 @@ module.exports = {
   snakeToCamel,
   isConstant,
   fileExt,
-  parsePath,
   trimLines,
 };
